requirements.txt
fastapi
uvicorn
loguru
aiofiles
python-multipart
python-docx
chardet
pandas
openpyxl
python-pptx
pdfplumber
pypdf
tabula-py
pdf2image
paddleocr
paddlespeech
moviepy